requirements.txt
pandas>=1.5.0
numpy>=1.21.0
gymnasium>=0.28.0
stable-baselines3[extra]>=2.0.0
torch>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0